+++ src/components/Breadcrumb/Breadcrumbs.stories.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import { Breadcrumb } from '../index';
+
+export default {
+  title: 'components/Breadcrumb',
+  component: Breadcrumb,
+  argTypes: {},
+};
+
+const data = [
+  {
+    title: 'Home',
+    href: '#'
+  },
+  {
+    title: 'Library',
+    href: '#'
+  },
+  {
+    title: 'Data',
+    href: '#'
+  }
+]
+
+const Template = (args) => {
+  return <Breadcrumb {...args} />;
+};
+
+export const Breadcrumb_Story = Template.bind({});
+Breadcrumb_Story.args = {
+  data
+}