+++ src/components/Form/index.js
@@ -1,0 +1,6 @@
+import { Form as BForm } from 'react-bootstrap';
+
+const Form = BForm;
+
+export default Form;
+export { Form };